+++ HISTORY.md
@@ -1,6 +1,11 @@
 # History of versions #
 
-* Version 1.3.0 (2019-01-06)
+* Version 1.4.1 (2019-03-26)
+  * Fixed COPY keybinding
+* Version 1.4.0 (2019-02-28)
+  * Added keybindings for COPY and PASTE
+  * Added keybindings to move focus between terminals in the same tab
+* Version 1.3.0 (2019-02-06)
   * Added the -e and -x parameters to create a new window an launch a command inside
   * Added the --working-directory parameter
   * Fixed the .desktop file to fully follow the standard